+++ src/main/java/com/taskone/restapi/model/EmployeeRequest.java
@@ -14,6 +14,6 @@ public class EmployeeRequest {
     private String username;
     private String email;
     private String jobposition;
-    private Integer salary;
+    private double salary;
 
 }
