+++ src/main/java/com/taskone/restapi/model/EmployeeResponse.java
@@ -17,6 +17,7 @@ public class EmployeeResponse {
     private String username;
     private String email;
     private String jobposition;
-    private Integer salary;
+    private double salary;
+
 
 }
